Add answer to Q52: CORS middleware
The documentation says: "CorsMiddleware should be placed as high as possible, especially before any middleware that can generate responses such as Django’s CommonMiddleware or Whitenoise’s WhiteNoiseMiddleware. If it is not before, it will not be able to add the CORS headers to these responses.". Talking about middleware config list on settings.py file